requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.3
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
rich==14.0.0
Werkzeug==3.1.3
